Initialize component PresetSelect
