+++ pages/index.js
@@ -3,7 +3,7 @@ import CardList from "@/components/CardList";
 export default function HomePage() {
   return (
     <>
-      {/* <a href="/test">Test Page</a> */}
+      <a href="/test">Test Page</a>
       <CardList />
     </>
   );
